dressy/requirements.txt
-e git+https://github.com/django/django.git@e0a98e2374b1d7660ef3841a5bb2626f1c69970a#egg=Django-master
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
SQLAlchemy==0.8.4
Tempita==0.5.2
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
astroid==1.0.1
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
django-allauth==0.18.0
dnspython==1.11.1
duplicity==0.6.23
html5lib==0.999
httplib2==0.8
lockfile==0.8
logilab-common==0.61.0
lxml==3.3.3
mutagen==1.22
nose==1.3.1
oauthlib==0.6.1
oneconf==0.3.7
pbr==0.7.0
pep8==1.4.6
pexpect==3.1
piston-mini-client==0.7.5
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pylint==1.1.0
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-dateutil==1.5
python-debian==0.1.21-nmu2ubuntu2
python-mpd==0.3.0
python-openid==2.2.5
pyxdg==0.25
reportlab==3.0
requests==2.2.1
requests-oauthlib==0.4.1
sessioninstaller==0.0.0
six==1.5.2
software-center-aptd-plugins==0.0.0
sqlalchemy-migrate==0.8.2
system-service==0.1.6
unity-lens-photos==1.0
urllib3==1.7.1
virtualenv==1.11.4
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose==3.6.3build2
zope.interface==4.0.5
